+++ Sources/FWFramework/Kernel/Router.swift
@@ -57,18 +57,39 @@ public class Router: NSObject, @unchecked Sendable {
         /// 路由是否以openURL方式打开，区别于objectForURL
         public fileprivate(set) var isOpening: Bool = false
 
+        /// 路由是否已处理完成，complete调用时自动设置
+        public private(set) var isCompleted: Bool = false
+
         /// 创建路由参数对象
         public init(url: String, userInfo: [AnyHashable: Any]? = nil, completion: Completion? = nil) {
             self.url = url
             self.userInfo = userInfo ?? [:]
             self.completion = completion
         }
+
+        /// 完成路由并回调结果
+        @MainActor public func complete(_ result: Sendable? = nil) {
+            isCompleted = true
+            completion?(result)
+        }
+
+        deinit {
+            if let routerCompleted = userInfo[Router.Parameter.routerCompletedKey] as? Bool,
+               routerCompleted, isOpening, !isCompleted {
+                isCompleted = true
+                if let completion {
+                    DispatchQueue.fw.mainDeinit {
+                        completion(nil)
+                    }
+                }
+            }
+        }
     }
 
     /// 路由处理句柄，仅支持openURL时可返回nil
     public typealias Handler = @MainActor @Sendable (Context) -> Any?
     /// 路由完成回调句柄
-    public typealias Completion = @MainActor @Sendable (Any?) -> Void
+    public typealias Completion = @MainActor @Sendable (Sendable?) -> Void
 
     /// 路由参数类，可直接使用，也可完全自定义
     open class Parameter: ObjectParameter, @unchecked Sendable {
@@ -78,6 +99,8 @@ public class Router: NSObject, @unchecked Sendable {
         public static let routerOptionsKey = "routerOptions"
         /// 路由动画选项Key，兼容字典传参，仅open生效
         public static let routerAnimatedKey = "routerAnimated"
+        /// 路由完成选项Key，兼容字典传参，仅open生效
+        public static let routerCompletedKey = "routerCompleted"
         /// 路由信息句柄Key，兼容字典传参，仅open生效
         public static let routerHandlerKey = "routerHandler"
 
@@ -87,6 +110,8 @@ public class Router: NSObject, @unchecked Sendable {
         open var routerOptions: NavigatorOptions?
         /// 路由动画选项，仅open生效
         open var routerAnimated: Bool?
+        /// 路由完成选项，仅open生效
+        open var routerCompleted: Bool?
         /// 路由信息句柄，仅open生效
         open var routerHandler: (@convention(block) @MainActor @Sendable (Context, UIViewController) -> Void)?
 
@@ -98,6 +123,7 @@ public class Router: NSObject, @unchecked Sendable {
                 self.routerOptions = options as? NavigatorOptions ?? NavigatorOptions(rawValue: NSNumber.fw.safeNumber(options).intValue)
             }
             self.routerAnimated = dictionaryValue[Self.routerAnimatedKey].bool
+            self.routerCompleted = dictionaryValue[Self.routerCompletedKey].bool
             self.routerHandler = dictionaryValue[Self.routerHandlerKey] as? @convention(block) @MainActor @Sendable (Context, UIViewController) -> Void
         }
 
@@ -106,6 +132,7 @@ public class Router: NSObject, @unchecked Sendable {
             dictionary[Self.routerSourceKey] = routerSource
             dictionary[Self.routerOptionsKey] = routerOptions
             dictionary[Self.routerAnimatedKey] = routerAnimated
+            dictionary[Self.routerCompletedKey] = routerCompleted
             dictionary[Self.routerHandlerKey] = routerHandler
             return dictionary
         }
@@ -283,42 +310,22 @@ public class Router: NSObject, @unchecked Sendable {
         return URLParameters[routeBlockKey] != nil
     }
 
-    /// 打开此 URL，带上附加信息，同时当操作完成时，执行额外的代码
+    /// 打开指定 URL，带上附加信息，同时当操作完成时，执行额外的代码
     /// - Parameters:
     ///   - url: 带 Scheme 的 URL，如 app://beauty/4
     ///   - userInfo: 附加信息
     ///   - completion: URL 处理完成后的 callback，完成的判定跟具体的业务相关
-    @MainActor public class func openURL(_ url: StringParameter?, userInfo: [AnyHashable: Any]? = nil, completion: Completion? = nil) {
-        let rewriteURL = rewriteURL(url)
-        guard !rewriteURL.isEmpty else { return }
-
-        let urlParameters = routeParameters(from: rewriteURL)
-        let handler = urlParameters[routeBlockKey] as? Handler
-        urlParameters.removeObject(forKey: routeBlockKey)
-
-        let context = Context(url: rewriteURL, userInfo: userInfo, completion: completion)
-        context.urlParameters = urlParameters as! [AnyHashable: Any]
-        context.isOpening = true
-
-        if let routeFilter {
-            if !routeFilter(context) { return }
-        }
-        if let handler {
-            let object = handler(context)
-            if let object, let routeHandler {
-                _ = routeHandler(context, object)
-            }
-            return
-        }
-        errorHandler?(context)
+    @discardableResult
+    @MainActor public class func openURL(_ url: StringParameter?, userInfo: [AnyHashable: Any]? = nil, completion: Completion? = nil) -> Any? {
+        openURL(url, userInfo: userInfo, isOpening: true, completion: completion)
     }
 
     /// 快速调用Handler参数中的回调句柄，指定回调结果
     /// - Parameters:
     ///   - context: Handler中的模型参数
     ///   - result: URL处理完成后的回调结果
-    @MainActor public class func completeURL(_ context: Context, result: Any?) {
-        context.completion?(result)
+    @MainActor public class func completeURL(_ context: Context, result: Sendable?) {
+        context.complete(result)
     }
 
     /// 查找谁对某个 URL 感兴趣，如果有的话，返回一个 object；如果没有，返回nil
@@ -327,29 +334,7 @@ public class Router: NSObject, @unchecked Sendable {
     ///   - userInfo: 附加信息
     /// - Returns: URL返回的对象
     @MainActor public class func object(forURL url: StringParameter?, userInfo: [AnyHashable: Any]? = nil) -> Any? {
-        let rewriteURL = rewriteURL(url)
-        guard !rewriteURL.isEmpty else { return nil }
-
-        let urlParameters = routeParameters(from: rewriteURL)
-        let handler = urlParameters[routeBlockKey] as? Handler
-        urlParameters.removeObject(forKey: routeBlockKey)
-
-        let context = Context(url: rewriteURL, userInfo: userInfo, completion: nil)
-        context.urlParameters = urlParameters as! [AnyHashable: Any]
-        context.isOpening = false
-
-        if let routeFilter {
-            if !routeFilter(context) { return nil }
-        }
-        if let handler {
-            let object = handler(context)
-            if let object, let routeHandler {
-                return routeHandler(context, object)
-            }
-            return object
-        }
-        errorHandler?(context)
-        return nil
+        openURL(url, userInfo: userInfo, isOpening: false, completion: nil)
     }
 
     /// 调用此方法来拼接 pattern 和 parameters
@@ -539,6 +524,32 @@ public class Router: NSObject, @unchecked Sendable {
             subRoutes = subRoutes[pathComponent] as? NSMutableDictionary ?? NSMutableDictionary()
         }
         return subRoutes
+    }
+
+    @MainActor private class func openURL(_ url: StringParameter?, userInfo: [AnyHashable: Any]?, isOpening: Bool, completion: Completion?) -> Any? {
+        let rewriteURL = rewriteURL(url)
+        guard !rewriteURL.isEmpty else { return nil }
+
+        let urlParameters = routeParameters(from: rewriteURL)
+        let handler = urlParameters[routeBlockKey] as? Handler
+        urlParameters.removeObject(forKey: routeBlockKey)
+
+        let context = Context(url: rewriteURL, userInfo: userInfo, completion: completion)
+        context.urlParameters = urlParameters as! [AnyHashable: Any]
+        context.isOpening = isOpening
+
+        if let routeFilter {
+            if !routeFilter(context) { return nil }
+        }
+        if let handler {
+            let object = handler(context)
+            if let object, let routeHandler {
+                return routeHandler(context, object)
+            }
+            return object
+        }
+        errorHandler?(context)
+        return nil
     }
 
     private class func pathComponents(from url: String) -> [String] {
@@ -820,5 +831,20 @@ extension Router {
             convertValue = originalValue.removingPercentEncoding ?? ""
         }
         return convertValue
+    }
+}
+
+// MARK: - Concurrency+Router
+extension Router {
+    /// 协程方式打开指定 URL，带上附加信息。业务侧需调用 completeURL 指定回调结果，如果未调用则会在Context释放时自动回调nil
+    @discardableResult
+    @MainActor public class func openURL(_ url: StringParameter?, userInfo: [AnyHashable: Any]? = nil) async -> Sendable? {
+        await withCheckedContinuation { continuation in
+            var userInfo = userInfo ?? [:]
+            userInfo[Router.Parameter.routerCompletedKey] = true
+            openURL(url, userInfo: userInfo) { result in
+                continuation.resume(returning: result)
+            }
+        }
     }
 }